refact: UserNames 클래스의 toString을 Formatter 클래스에서 Format에 맞춰서 반환하도록 수정
